+++ CHANGELOG.md
@@ -1,7 +1,9 @@
-# Current
+# 0.25.3
  
+- Add: Unichain
 - Add: `Client.fetch_token_metadata`: Load token metadata for given token addresses, including Coingecko and TokenSniffer data
 - Add: `token_extra_data.load_token_metadata`: Load and cache token metadata for given DataFrame of trading pairs
+- Add: `token_filter.filter_by_token_sniffer_score`: Load and cache token metadata for given DataFrame of trading pairs
 
 # 0.25.2